Test SpotifyRemoteServer listening to socket events
This adds test for SpotifyRemoteServer's behaviour regarding sockets event and
asserts what SpotifyRemoteServer does when such an event is triggered.

I had to add some simple `on` and `trigger` stubs to the socket stub in the test
suite. Also, the spotify methods that get called on certain events are stubbed
in each test.

The tests assert that SpotifyRemoteServer really reacts to socket events by
emitting the current state or current track.

I also changed the test suite to use `server.stopIntervals()` in each test
(after the assertions) so they don't interfere with other tests making
assertions about the call count of certain methods that get called in intervals.