Add mute/unmute button to UI
This commit adds a mute/unmute button to the UI, which makes use of the
`muteVolume` and `unmuteVolume` functions provided by node-spotify-applescript
module.

`SpotifyRemoteServer` wraps these functions and sends the new state to the
clients after successfully muting/unmuting the volume. I added tests for this
behaviour.

The client adjusts its icon/button depending on the state sent by the server.
It also remembers whether the volume has been changed by using the range input
or by clicking on the mute/unmute button.